+++ src/main/java/net/sf/jabref/logic/importer/Parser.java
@@ -5,6 +5,9 @@ import java.util.List;
 
 import net.sf.jabref.model.entry.BibEntry;
 
+/**
+ * A parser converts an {@link InputStream} into a list of {@link BibEntry}.
+ */
 public interface Parser {
 
     List<BibEntry> parseEntries(InputStream inputStream) throws ParserException;